Added password validator
