+++ backend/itm_backend/api/serializers.py
@@ -1,15 +1,34 @@
+import base64
+
 from django.contrib.auth import get_user_model
 from django.contrib.auth.hashers import make_password
+from django.core.files.base import ContentFile
+from projects.models import Project, Tag, Task
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
-
 from users.models import TimeZone
-from projects.models import Project, Task, Tag
+
+from .validators import validate_password
 
 User = get_user_model()
 
 
 OFFSET_RANGE = (-12, 15)
+
+
+class Base64ImageField(serializers.ImageField):
+    """Сериализация и десериализация изображений в формат base64."""
+
+    def to_internal_value(self, data):
+        """Преобразует изображение в формате base64 в объект ContentFile Django."""
+
+        if isinstance(data, str) and data.startswith('data:image'):
+            format, imgstr = data.split(';base64,')
+            ext = format.split('/')[-1]
+
+            data = ContentFile(base64.b64decode(imgstr), name='temp.' + ext)
+
+        return super().to_internal_value(data)
 
 
 class TimeZoneSerializer(serializers.HyperlinkedModelSerializer):
@@ -39,6 +58,8 @@ class CustomUserCreateSerializer(serializers.ModelSerializer):
     Здесь определены поля, которые будут отображаться при создании пользователя.
     """
 
+    password = serializers.CharField(validators=[validate_password])
+
     class Meta:
         model = User
         fields = ["email", "password", "first_name", "last_name"]
@@ -49,6 +70,17 @@ class CustomUserCreateSerializer(serializers.ModelSerializer):
         """
         validated_data["password"] = make_password(validated_data["password"])
         return super().create(validated_data)
+
+    def to_representation(self, instance):
+        """
+        Возвращает информацию о пользователе по ТЗ.
+        """
+        return {
+            "id": instance.id,
+            "email": instance.email,
+            "first_name": instance.first_name,
+            "last_name": instance.last_name
+        }
 
 
 class CustomUserSerializer(serializers.ModelSerializer):
@@ -62,6 +94,7 @@ class CustomUserSerializer(serializers.ModelSerializer):
     """
 
     timezone = TimeZoneSerializer()
+    photo = Base64ImageField()
 
     class Meta:
         """
